+++ src/main/java/online/kingdomkeys/kingdomkeys/effects/StopEffect.java
@@ -4,6 +4,7 @@ import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectCategory;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.player.Player;
 
 public class StopEffect extends MobEffect {
     public StopEffect(MobEffectCategory pCategory, int pColor) {
@@ -12,6 +13,10 @@ public class StopEffect extends MobEffect {
 
     @Override
     public boolean applyEffectTick(LivingEntity pLivingEntity, int pAmplifier) {
+        if (pLivingEntity instanceof Player player) {
+            if (player.isCreative())
+                return false;
+        }
         pLivingEntity.setDeltaMovement(0, 0, 0);
         pLivingEntity.hurtMarked = true;
 
@@ -21,7 +26,6 @@ public class StopEffect extends MobEffect {
 
         return super.applyEffectTick(pLivingEntity, pAmplifier);
     }
-
 
     public boolean shouldApplyEffectTickThisTick(int pDuration, int pAmplifier) {
         return true;