+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 **Changed:**
 * Stop will now deal the damage gradually rather than all at once.
+* Stop and Gravity will no longer work with players in creative.
 
 **Fixed:**
 * Stop dealing twice the damage than it should.